- Changing text size for log - Adding id for log text view - Adding padding for log text view - Setting remainingMinutes class variable - Adding text constants for break switch button - Adding class variable for employee status - Adding click listener for break and cancel button - Setting initial text - Changing text of break button when pressed and updating employee status - Finishing log when clicking on cancel button

+++ app/src/main/java/com/example/workintime/MainActivity.kt
@@ -3,7 +3,6 @@ package com.example.workintime
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import com.example.workintime.databinding.ActivityMainBinding
 
 val hoursParamId = "com.example.workintime.hours"
@@ -33,7 +32,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun startDay(){
-        val hours = binding.initialHoursSlider.value
+        val hours = binding.initialHoursSlider.value.toInt()
         val intent = Intent(this, LogActivity::class.java).apply {
             putExtra(hoursParamId, hours)
         }
